requirements.txt
torch
torchvision
numpy
pandas
matplotlib
scikit-learn
pillow
tqdm
PyYAML
requests
kaggle
streamlit